Move utilities file and update import paths
