+++ src/components/MovieDetails/MovieDetails.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import "./MovieDetails.css";
 import Description from "../Description/Description.js";
 import Image from "../Image/Image.js";
-import { scrubDetailsData } from "../utilities.js";
+import { scrubDetailsData } from "../../utilities.js";
 import { testAPI } from "../../api.js"
 
 class MovieDetails extends Component {
